attempt to switch to slidev
I made use of the affordances of the old system (marp) and this one is
looking real bad. I would have made it differently if I had used this
from the beginning, but switching it over is burning too much time. I
think it's time to try a new way.